% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{d.s}
\alias{d.s}
\title{The diamond-step algorithm}
\usage{
d.s(mat, sd)
}
\arguments{
\item{mat}{This inputs the matrix on which the diamond-step algorithm will act upon}

\item{sd}{This is the noise that will be added to each iteration of the DS algorithm}
}
\value{
returns the matrix, now manipulated to have a center
}
\description{
The diamond-step algorithm
}

