% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{dss}
\alias{dss}
\title{The diamond-square-step algorithm}
\usage{
dss(n, sd)
}
\arguments{
\item{n}{Will create the matrix}

\item{sd}{This is the noise that will be added to each iteration of the DSS algorithm}
}
\value{
returns the matrix, now manipulated by the DSS algorithm
}
\description{
The diamond-square-step algorithm
}

