% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{eco.gen}
\alias{eco.gen}
\title{Setting up plant information}
\usage{
eco.gen(terrain, timesteps)
}
\arguments{
\item{terrain}{Give original 2D terrain matrix}

\item{timesteps}{Inputs how many timesteps will be present in ecosystem}
}
\value{
returns an array that matches matrix at some number of timesteps
}
\description{
Setting up plant information
}

