% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{genmat}
\alias{genmat}
\title{Generates starting terrain matrix; includes starting values}
\usage{
genmat(n)
}
\arguments{
\item{n}{Size of the matrix will be (2^n)+1, which will always produce a
matrix that is subdivisible into matrices with odd sides}
}
\value{
a terrain matrix
}
\description{
Generates starting terrain matrix; includes starting values
}

