% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Plants.r
\name{make.plants}
\alias{make.plants}
\title{Generate plants to exist, reproduce, and compete within the terrain}
\usage{
make.plants(terrain, survive = c(0.7, 0.85), repro = c(0.95, 0.55),
  names = NULL, timesteps = 50)
}
\arguments{
\item{survive}{The rate of survival of the listed plants. (default: .8, .65)
Is a vector of length equal to the number of listed plants}

\item{repro}{The rate of reproduction of the listed plants. (default: .4,.8)
Is a vector of length equal to the number of listed plants}

\item{names}{The names of listed plants. (defaul=NULL)
If no names are provided, plants will be assigned a letter in the alphabet 
Is a vector of length equal to the number of listed plants}

\item{timesteps}{The number of 'turns' that you want the simulation to exectute. (defaul: 50)}
}
\value{
a plant matrix; empty cells contain only '', waterlogged cells contain Na,  
         and cells containing plants will have the indicated (or assigned) name.
}
\description{
Simulates plants on a terrain which is built in
terrain.r. Plants survive and reproduce in terrain that has a height 0 <= 
Ensure that you have the same length of vectors for params repro, survive and names!!
}
\examples{
plants <- make.plants(terrain, c(.7,.85), repro=c(.95,.55), names=NULL, 50)
}
