% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Terrain.r
\name{make.terrain}
\alias{make.terrain}
\title{Terrain with elivations represented by numerics in each cell}
\usage{
make.terrain(n = 6, water = TRUE, sd = 15)
}
\arguments{
\item{n}{Determines the size of the grid. Grid will be 2^n +1 with a default of n=6}

\item{water}{a logical to indicate whether terrain lower than 0 should be underwater (default:TRUE)}

\item{sd}{the random noise to be added to each step of diamond.square.step}
}
\value{
a terrain matrix; numeric elements indicate height, NAs indicate cells filled with water
}
\description{
Makes a matrix which is filled according to the diamond step algorithm.
Altitude is represented by values in each cell. Cells with NA are waterlogged.

This is a wrapper around \code{diamond.square.step} and \code{add.water}
}
\examples{
terry <- make.terrain(4, 15)
image(terry)

}
