% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{plant.timestep}
\alias{plant.timestep}
\title{Timestep function of plants in ecosystem}
\usage{
plant.timestep(eco, info)
}
\arguments{
\item{eco}{The ecosystem array containing information on plant position at each time step}

\item{info}{Info is a list contianing all of the parameter information of the ecosystem}
}
\value{
returns the ecosystem array, now with plants at all timesteps having reproduced in adjacent cells
}
\description{
Timestep function of plants in ecosystem
}

