% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{reproduce}
\alias{reproduce}
\title{reproduction function of plants in ecosystem}
\usage{
reproduce(row, col, time, eco, info)
}
\arguments{
\item{row}{The row position of the ecosystem array to be input}

\item{col}{The column position of the ecosystem array to be input}

\item{time}{Which timestep in the array that will be input}

\item{eco}{The ecosystem array}

\item{info}{Info is a list contianing all of the parameter information of the ecosystem}
}
\value{
returns the ecosystem array, now with plants having reproduced and undergone competition from surrounding cells
}
\description{
reproduction function of plants in ecosystem
}

