% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{run.plant.ecosystem}
\alias{run.plant.ecosystem}
\title{Make an time series plant ecosystem with water}
\usage{
run.plant.ecosystem(terrain, timesteps, survival, reproduction, competition)
}
\arguments{
\item{terrain}{dlafkdjfaklj}

\item{timesteps}{dlfajdkfj}

\item{survival}{ldkfajdlkjf}

\item{reproduction}{dlfadkfja}

\item{competition}{ldkfjadklj}
}
\value{
a time series plant/water composition array; character elements including single letters
indicate presence of a corresponding plant species while NA indicates water, "" indicates absolute absence
}
\description{
A light wrapper around \code{setup.plants} and \code{plant.timestep}
}
\examples{
rworld <- run.plant.ecosystem(terrain = make.terrain(4), timesteps=100, survival=c(.5,.5), reproduction=c(.5,.5), competition=matrix(c(.50,.50,.50,.50), 2))
}

