% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{s.s}
\alias{s.s}
\title{The square-step algorithm}
\usage{
s.s(mat, sd)
}
\arguments{
\item{mat}{This inputs the matrix on which the square-step algorithm will act upon}

\item{sd}{This is the noise that will be added to each iteration of the SS algorithm}
}
\value{
returns the matrix, now manipulated to contain north/east/south/west points
   originating from the center produced by the diamond-step
}
\description{
The square-step algorithm
}

