% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{setup.plants}
\alias{setup.plants}
\title{Setting up plant information}
\usage{
setup.plants(names = NULL)
}
\arguments{
\item{names}{a parameter of NULL}
}
\value{
returns a list containing all the parameters that will constitute the ecosystem
}
\description{
Setting up plant information
}

