% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plants.R
\name{survive}
\alias{survive}
\title{Survival function of plants in ecosystem}
\usage{
survive(cell, info)
}
\arguments{
\item{cell}{This is a particular cell in the ecosystem at a particular time step}

\item{info}{Info is a list contianing all of the parameter information of the ecosystem}
}
\value{
returns the cell to be input back into the larger ecosystem matrix
}
\description{
Survival function of plants in ecosystem
}

