% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{terrain.gen}
\alias{terrain.gen}
\title{The wrapper function for the diamond-square-step algorithm}
\usage{
terrain.gen(n, sd, lakes)
}
\arguments{
\item{n}{Generates original matrix with side lengths (2^n)+1}

\item{sd}{This is the noise that will be added to each iteration of the DSS algorithm}

\item{lakes}{If True, will induce values in the matrix less than 0 to be NA values, indicating water}
}
\value{
returns the completed matrix
}
\description{
The wrapper function for the diamond-square-step algorithm
}

