% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/terrain.R
\name{make.terrain}
\alias{make.terrain}
\title{Make an elevational grid with water}
\usage{
make.terrain(n)
}
\arguments{
\item{n}{Size of grid will be a 2^n +1 grid (default: 4; a 17 x 17 grid)
diamond.square algorithm; inputs initially constructed terrain matrix and performs
diamond square steps on all cells}
}
\value{
a terrain matrix; numeric elements indicate heights, and
NAs indicate cells filled with water

an image; a grid with colors corresponding to the values in the terrain matrix
}
\description{
A light wrapper around \code{diamond.square.step}
}
\examples{
terrain <- make.terrain(4)
}

